requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==40.0.2
decorator==5.1.1
Django==4.1.6
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-widget-tweaks==1.4.12
docopt==0.6.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
pipreqs==0.4.11
pycparser==2.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.14
yarg==0.1.9
pillow==10.3.0
django-ckeditor